Update which sections get shown
We've implemented some more aggressive updating procedures which is
leading to more clashes with the README when multiple module update PRs
are opened.

I don't think we're interested in modules and resources, they're
implementation details really, so I've updated the configuration to only
document the module's interfaces: custom documentation, requirements,
variables, and outputs.